Add TODO for adding group variables
At the moment, have to set this manually which is a bit of a pain. Need to try and make it so that when/if we generate
the keys we add the server group config to specify which key to use.

Also, same for username depending on whether it's DO or AWS.

Issue(s): None